Trait democracy::pallet::Config

pub trait Config: Config + Sized {

    type Proposal: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + From<Call<Self>>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type EnactmentPeriod: Get<Self::BlockNumber>;
    type LaunchPeriod: Get<Self::BlockNumber>;
    type VotingPeriod: Get<Self::BlockNumber>;
    type MinimumDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type FastTrackOrigin: EnsureOrigin<Self::Origin>;
    type FastTrackVotingPeriod: Get<Self::BlockNumber>;
    type PreimageByteDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type Scheduler: ScheduleNamed<Self::BlockNumber, Self::Proposal, Self::PalletsOrigin>;
    type PalletsOrigin: From<RawOrigin<Self::AccountId>>;
    type MaxVotes: Get<u32>;
    type WeightInfo: WeightInfo;
    type MaxProposals: Get<u32>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Proposal: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + From<Call<Self>>

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Currency: ReservableCurrency<Self::AccountId>

Currency type for this pallet.

type EnactmentPeriod: Get<Self::BlockNumber>

The period between a proposal being approved and enacted.

It should generally be a little more than the unstake period to ensure that voting stakers have an opportunity to remove themselves from the system in the case where they are on the losing side of a vote.

type LaunchPeriod: Get<Self::BlockNumber>

How often (in blocks) new public referenda are launched.

type VotingPeriod: Get<Self::BlockNumber>

How often (in blocks) to check for new votes.

type MinimumDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The minimum amount to be used as a deposit for a public referendum proposal.

type FastTrackOrigin: EnsureOrigin<Self::Origin>

Origin from which the next majority-carries (or more permissive) referendum may be tabled to vote according to the FastTrackVotingPeriod asynchronously in a similar manner to the emergency origin. It retains its threshold method.

type FastTrackVotingPeriod: Get<Self::BlockNumber>

Minimum voting period allowed for a fast-track referendum.

type PreimageByteDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount of balance that must be deposited per byte of preimage stored.

type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

Handler for the unbalanced reduction when slashing a preimage deposit.

type Scheduler: ScheduleNamed<Self::BlockNumber, Self::Proposal, Self::PalletsOrigin>

The Scheduler.

type PalletsOrigin: From<RawOrigin<Self::AccountId>>

Overarching type of all pallets origins.

type MaxVotes: Get<u32>

The maximum number of votes for an account.

Also used to compute weight, an overly big value can lead to extrinsic with very big weight.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type MaxProposals: Get<u32>

The maximum number of public proposals that can exist at any time.